% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_bd}
\alias{read_bd}
\alias{read_bd.data.frame}
\alias{read_bd.character}
\alias{read_bd_as_df}
\title{read_bd}
\usage{
read_bd(...)

\method{read_bd}{data.frame}(x, split = FALSE)

\method{read_bd}{character}(filename, samplename = NULL)

read_bd_as_df(filename, samplename = NULL)
}
\arguments{
\item{filename}{name of file to read.}

\item{samplename}{sample names to attach to cell_index.}
}
\description{
Read file in BD format.
}
